+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -7,11 +7,11 @@ import java.util.Collection;
 public interface FilmStorage {
     public Collection<Film> list();
 
-    public Film get(int filmId);
+    public Film get(long filmId);
 
     public Film create(Film film);
 
     public Film update(Film film);
 
-    public boolean notExists(int filmId);
+    public boolean notExists(long filmId);
 }
